split builder
