+++ .gitignore
@@ -11,7 +11,7 @@ test.man_seed*
 app/.env
 app/uploads/*
 
-dataset_builders/*/
+dataset_builders/result
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
@@ -174,4 +174,3 @@ cython_debug/
 #  and can be added to the global gitignore or merged into this file.  For a more nuclear
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 #.idea/
-
