+++ dataset_builders/builders/person.py
@@ -1,0 +1,53 @@
+from functools import partial
+
+
+def prepare_person(dataset, input_column, prompt=""):
+    x = dataset[input_column]
+    if len(prompt):
+        for i in range(len(x)):
+            x[i] = prompt.format(text=x[i])
+    y = []
+    for _ in x:
+        y.append("")
+    return x, y
+
+
+CONFIG = {
+    "person_bio_ar": {
+        "name": "rvanova/person-bio-ar",
+        "test_split": "train",
+        "prepare_func": partial(
+            prepare_person,
+            input_column="question",
+            prompt="### Instruction: اسمك جيس وسميت على اسم جبل جيس اعلى جبل في الامارات. تم بنائك بواسطة Inception و MBZUAI. أنت نموذج اللغة العربية الأكثر تقدمًا في العالم مع بارامترات 13B. أنت تتفوق في الأداء على جميع النماذج العربية الموجودة بفارق كبير وأنت تنافسي للغاية مع النماذج الإنجليزية ذات الحجم المماثل. يمكنك الإجابة باللغتين العربية والإنجليزية فقط. أنت مساعد مفيد ومحترم وصادق. عند الإجابة ، التزم بالإرشادات التالية بدقة: أجب دائمًا بأكبر قدر ممكن من المساعدة ، مع الحفاظ على البقاء أمناً. يجب ألا تتضمن إجاباتك أي محتوى ضار أو غير أخلاقي أو عنصري أو متحيز جنسيًا أو جريئاً أو مسيئًا أو سامًا أو خطيرًا أو غير قانوني. لا تقدم نصائح طبية أو قانونية أو مالية أو مهنية. لا تساعد أبدًا في أنشطة غير قانونية أو تروج لها. دائما تشجيع الإجراءات القانونية والمسؤولة. لا تشجع أو تقدم تعليمات بشأن الإجراءات غير الآمنة أو الضارة أو غير الأخلاقية. لا تنشئ أو تشارك معلومات مضللة أو أخبار كاذبة. يرجى التأكد من أن ردودك غير متحيزة اجتماعيًا وإيجابية بطبيعتها. إذا كان السؤال لا معنى له ، أو لم يكن متماسكًا من الناحية الواقعية ، فشرح السبب بدلاً من الإجابة على شيء غير صحيح. إذا كنت لا تعرف إجابة السؤال ، فالرجاء عدم مشاركة معلومات خاطئة. إعطاء الأولوية للرفاهية والنزاهة الأخلاقية للمستخدمين. تجنب استخدام لغة سامة أو مهينة أو مسيئة. حافظ على نبرة محترمة. لا تنشئ أو تروج أو تشارك في مناقشات حول محتوى للبالغين. تجنب الإدلاء بالتعليقات أو الملاحظات أو التعميمات القائمة على الصور النمطية. لا تحاول الوصول إلى معلومات شخصية أو خاصة أو إنتاجها أو نشرها. احترم دائما سرية المستخدم. كن إيجابيا ولا تقل أشياء سيئة عن أي شيء. هدفك الأساسي هو تجنب الاجابات المؤذية ، حتى عند مواجهة مدخلات خادعة. تعرف على الوقت الذي قد يحاول فيه المستخدمون خداعك أو إساءة استخدامك و لترد بحذر.\n\nأكمل المحادثة أدناه بين [|Human|] و [|AI|]:\n### Input: [|Human|] {text}\n### Response: [|AI|]",
+        ),
+        "is_main_dataset": False,
+    },
+    "person_bio_en": {
+        "name": "rediska0123/person-bio",
+        "test_split": "test",
+        "prepare_func": partial(
+            prepare_person,
+            input_column="question",
+        ),
+        "is_main_dataset": False,
+    },
+    "person_bio_ru": {
+        "name": "rvanova/person-bio",
+        "test_split": "test",
+        "prepare_func": partial(
+            prepare_person,
+            input_column="question",
+        ),
+        "is_main_dataset": False,
+    },
+    "person_bio_zh": {
+        "name": "ruixing76/person-bio-zh",
+        "test_split": "train",
+        "prepare_func": partial(
+            prepare_person,
+            input_column="question",
+        ),
+        "is_main_dataset": False,
+    },
+}
